move Image Comments Adapters into separate classes
